change from function to class components
