+++ packages/pipeline-editor/src/NotebookSubmissionDialog.tsx
@@ -17,19 +17,14 @@
 import { IDictionary } from '@elyra/application';
 import * as React from 'react';
 
-import { IRuntime } from './PipelineEditorWidget';
+import { IRuntime } from './PipelineService';
+import Utils from './utils';
 
-type Props = {
+interface IProps {
   runtimes: IRuntime[];
   images: IDictionary<string>;
   env: string[];
-};
-
-const chunkArray = <T extends {}>(arr: T[], n: number): T[][] => {
-  return Array.from(Array(Math.ceil(arr.length / n)), (_, i) =>
-    arr.slice(i * n, i * n + n)
-  );
-};
+}
 
 const EnvForm = ({ env }: { env: string[] }): JSX.Element => {
   if (env.length > 0) {
@@ -45,7 +40,7 @@ const EnvForm = ({ env }: { env: string[] }): JSX.Element => {
             </div>
           </td>
         </tr>
-        {chunkArray(env, 4).map((col, i) => (
+        {Utils.chunkArray(env, 4).map((col, i) => (
           <tr key={i}>
             {col.map(envVar => (
               <td key={envVar}>
@@ -68,77 +63,76 @@ const EnvForm = ({ env }: { env: string[] }): JSX.Element => {
   return null;
 };
 
-const NotebookSubmissionDialog = ({
-  runtimes,
-  images,
-  env
-}: Props): JSX.Element => {
-  return (
-    <form>
-      <table id="table-submit-dialog" className="elyra-table">
-        <tbody>
-          <tr>
-            <td colSpan={2}>
-              <label htmlFor="runtime_config">Runtime Config:</label>
-              <br />
-              <select
-                id="runtime_config"
-                name="runtime_config"
-                className="elyra-form-runtime-config"
-              >
-                {runtimes.map(runtime => (
-                  <option key={runtime.name} value={runtime.name}>
-                    {runtime.display_name}
-                  </option>
-                ))}
-              </select>
-            </td>
-            <td colSpan={2}>
-              <label htmlFor="framework">Runtime images:</label>
-              <br />
-              <select
-                id="framework"
-                name="framework"
-                className="elyra-form-framework"
-              >
-                {Object.entries(images).map(([key, val]) => (
-                  <option key={key} value={key}>
-                    {val}
-                  </option>
-                ))}
-              </select>
-            </td>
-          </tr>
-          <tr>
-            <td>
-              <br />
-              <input
-                type="checkbox"
-                id="dependency_include"
-                name="dependency_include"
-                size={20}
-                defaultChecked
-              />
-              <label htmlFor="dependency_include">Include dependencies</label>
-              <br />
-            </td>
-            <td colSpan={3}>
-              <br />
-              <input
-                type="text"
-                id="dependencies"
-                name="dependencies"
-                placeholder="*.py"
-                defaultValue="*.py"
-                size={20}
-              />
-            </td>
-          </tr>
-          <EnvForm env={env} />
-        </tbody>
-      </table>
-    </form>
-  );
-};
+class NotebookSubmissionDialog extends React.Component<IProps> {
+  render(): React.ReactNode {
+    const { runtimes, images, env } = this.props;
+    return (
+      <form>
+        <table id="table-submit-dialog" className="elyra-table">
+          <tbody>
+            <tr>
+              <td colSpan={2}>
+                <label htmlFor="runtime_config">Runtime Config:</label>
+                <br />
+                <select
+                  id="runtime_config"
+                  name="runtime_config"
+                  className="elyra-form-runtime-config"
+                >
+                  {runtimes.map(runtime => (
+                    <option key={runtime.name} value={runtime.name}>
+                      {runtime.display_name}
+                    </option>
+                  ))}
+                </select>
+              </td>
+              <td colSpan={2}>
+                <label htmlFor="framework">Runtime images:</label>
+                <br />
+                <select
+                  id="framework"
+                  name="framework"
+                  className="elyra-form-framework"
+                >
+                  {Object.entries(images).map(([key, val]) => (
+                    <option key={key} value={key}>
+                      {val}
+                    </option>
+                  ))}
+                </select>
+              </td>
+            </tr>
+            <tr>
+              <td>
+                <br />
+                <input
+                  type="checkbox"
+                  id="dependency_include"
+                  name="dependency_include"
+                  size={20}
+                  defaultChecked
+                />
+                <label htmlFor="dependency_include">Include dependencies</label>
+                <br />
+              </td>
+              <td colSpan={3}>
+                <br />
+                <input
+                  type="text"
+                  id="dependencies"
+                  name="dependencies"
+                  placeholder="*.py"
+                  defaultValue="*.py"
+                  size={20}
+                />
+              </td>
+            </tr>
+            <EnvForm env={env} />
+          </tbody>
+        </table>
+      </form>
+    );
+  }
+}
 
 export default NotebookSubmissionDialog;
