+++ packages/pipeline-editor/src/utils.ts
@@ -152,4 +152,17 @@ export default class Utils {
       this.deletePipelineAppdataField(node, currentFieldName);
     }
   }
+
+  /**
+   * Break an array into an array of "chunks", each "chunk" having "n" elements.
+   * The final "chuck" may have less than "n" elements.
+   * Example:
+   * chunkArray(['a', 'b', 'c', 'd', 'e', 'f', 'g'], 4)
+   * -> [['a', 'b', 'c', 'd'], ['e', 'f', 'g']]
+   */
+  static chunkArray<T extends {}>(arr: T[], n: number): T[][] {
+    return Array.from(Array(Math.ceil(arr.length / n)), (_, i) =>
+      arr.slice(i * n, i * n + n)
+    );
+  }
 }
